refactor(openclaw-plugin): reorganize into folder-per-surface, fix safety gaps
Split hooks/tools/cli into folder structure for scalability.
Add injection check on memory_store (defence in depth).
Add injection filter + HTML escape on memory_search results.
