--- packages/openclaw-memory-memwal/src/hooks.ts
@@ -1,133 +1,0 @@
-/**
- * Lifecycle hooks — the invisible memory layer.
- *
- * before_prompt_build: search MemWal for relevant memories, inject into prompt
- * agent_end: extract conversation text, send to MemWal for fact extraction
- *
- * Each agent gets its own namespace derived from ctx.sessionKey.
- * Same key, same account — isolation via server-side namespace scoping.
- */
-
-import type { MemWal } from "@cmdoss/memwal";
-import { resolveAgent } from "./config.js";
-import { shouldCapture, looksLikeInjection } from "./capture.js";
-import {
-  formatMemoriesForPrompt,
-  extractMessageTexts,
-  withRetry,
-} from "./format.js";
-import type { PluginConfig } from "./types.js";
-
-const MIN_PROMPT_LENGTH = 10;
-
-/** Register before_prompt_build (auto-recall) and agent_end (auto-capture) hooks. */
-export function registerHooks(api: any, client: MemWal, config: PluginConfig): void {
-  // Auto-recall: inject relevant memories before each agent turn
-  if (config.autoRecall) {
-    api.on("before_prompt_build", async (event: any, ctx: any) => {
-      // Skip trivial prompts ("ok", "y") — not worth a server round-trip
-      if (!event.prompt || event.prompt.length < MIN_PROMPT_LENGTH) return;
-
-      const { namespace, agentName } = resolveAgent(config.defaultNamespace, ctx?.sessionKey);
-
-      // The LLM doesn't know which agent it is. This instruction guides
-      // memory_search/memory_store tool calls to the correct namespace.
-      // Returned via appendSystemContext in ALL code paths (including errors)
-      // so tools still scope correctly even when recall itself fails.
-      const namespaceInstruction =
-        `When using memory_search or memory_store tools, ` +
-        `pass namespace="${namespace}" to scope operations to the current agent's memory.`;
-
-      try {
-        const result = await client.recall(
-          event.prompt,
-          config.maxRecallResults,
-          namespace,
-        );
-
-        if (!result.results?.length) {
-          return { appendSystemContext: namespaceInstruction };
-        }
-
-        // Two filters: relevance threshold (drop noise) and injection
-        // detection (drop memories containing prompt manipulation attempts)
-        const relevant = result.results.filter(
-          (r: any) =>
-            (1 - r.distance) >= config.minRelevance &&
-            !looksLikeInjection(r.text),
-        );
-
-        if (!relevant.length) {
-          return { appendSystemContext: namespaceInstruction };
-        }
-
-        api.logger.info(
-          `memory-memwal: auto-recall injected ${relevant.length} memories ` +
-          `(agent: ${agentName}, namespace: ${namespace})`,
-        );
-
-        return {
-          prependContext: formatMemoriesForPrompt(
-            relevant.map((r: any) => ({ text: r.text })),
-          ),
-          appendSystemContext: namespaceInstruction,
-        };
-      } catch (err) {
-        api.logger.warn(
-          `memory-memwal: auto-recall failed: ${String(err)}`,
-        );
-        return { appendSystemContext: namespaceInstruction };
-      }
-    });
-  }
-
-  // Auto-capture: extract and store facts after each agent turn
-  if (config.autoCapture) {
-    api.on("agent_end", async (event: any, ctx: any) => {
-      if (!event.success || !event.messages?.length) return;
-
-      const { namespace, agentName } = resolveAgent(config.defaultNamespace, ctx?.sessionKey);
-
-      try {
-        // Extract clean message texts (without role prefixes)
-        const texts = extractMessageTexts(
-          event.messages,
-          config.captureMaxMessages,
-        );
-
-        if (!texts.length) return;
-
-        // Filter individual messages — skip if none are worth capturing
-        const capturable = texts.filter((t) => shouldCapture(t));
-        if (!capturable.length) {
-          api.logger.debug?.(
-            `memory-memwal: auto-capture skipped — no capturable content ` +
-            `(agent: ${agentName}, ${texts.length} messages checked)`,
-          );
-          return;
-        }
-
-        // Numbered list helps the server LLM distinguish separate messages
-        // during fact extraction (vs one big wall of text)
-        const conversation = capturable
-          .map((t, i) => `${i + 1}. ${t}`)
-          .join("\n\n");
-
-        // analyze() calls the server LLM for fact extraction — retry once
-        // since transient failures are common with remote LLM calls
-        const result = await withRetry(() => client.analyze(conversation, namespace));
-
-        if (result.facts?.length) {
-          api.logger.info(
-            `memory-memwal: auto-captured ${result.facts.length} facts ` +
-            `(agent: ${agentName}, namespace: ${namespace})`,
-          );
-        }
-      } catch (err) {
-        api.logger.warn(
-          `memory-memwal: auto-capture failed: ${String(err)}`,
-        );
-      }
-    });
-  }
-}
